Return HarvestChoice indicator categories (3-level deep)

Return a compact list of indicator categories, codes and labels

category(q = NULL, as.list = FALSE)

Arguments

q

character array of pattern(s) to search for. If omitted will return all available indicators.

as.list

if TRUE returns a list of indicator codes grouped by category

Value

a data.table showing the number of indicators in each category, or a list of indicators grouped by category

Details

# API call: list all HarvestChoice indicators matching category 'demographics' $ curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/category/json \ -d '{"q" : "demographics' \ -X POST -H "Content-Type:application/json"

# API call: list all HarvestChoice indicators matching 'cassava' in a hierarchical list $ curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/category/json \ -d '"q" :" cassava", "as.list" : "true" \ -X POST -H "Content-Type:application/json"

# To return a complete list of published indicators omit 'q' $ curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/category/json \ -X POST -H "Content-Type:application/json"

See also

indicator to view complete indicator metadata records

Examples

# List all HarvestChoice indicators matching category 'demographics'
category("demographics")

#> Error in `[.data.frame`(x, i, j): could not find function "."

# List all HarvestChoice indicators matching 'cassava' in a hierarchical list
category("cassava", as.list=TRUE)

#> Error in `[.data.frame`(x, i, j): could not find function "."