classify {hcapi3}	R Documentation

Create custom domains for spatial targeting

Description

All HarvestChoice spatial layers come with default value breaks to create map legends and to automatically convert continuous variables into classified variables (this is needed when summarizing). For example, the value breaks for cassava production may be returned using getMeta("cass_p"). They are c(0, 1074, 4260, 10051, 19920, 40085, 101843). This method extends the classification logic to allow for any custom value break(s). For example one can summarize a set of selected indicators over areas of low population density, and/or areas covering a single farming system (or any combination of custom domains).

Usage

classify(var, by, ...)

Arguments

var
character array of variable names (all types are accepted)

by
named list of numeric or character arrays representing custom value breaks for each variable to summarize by

...
other arguments passed to getLayer() (e.g. iso3, collapse)

Value

a data.table of var indicators aggregated by custom by domains

Examples

# Return average wealth index and poverty rates for areas of medium rural
# population density and low market access (under 5hrs travel time)
classify(var=c("wealth","TPOV_P200"),
    by=list(PD05_RUR=c(60, 100), TT_20K=c(0,5)))

# Equivalent cUrl request at the command line
curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/classify/json \
-d '{"var" : ["wealth", "TPOV_P200"],
    "by" : {"PD05_RUR" : [60, 100], "TT_20K" : [0, 5]}}' \
-X POST -H "Content-Type:application/json"

[Package hcapi3 version 3.5 Index]