Query, subset, and/or summarize HarvestChoice indicators

Workhorse method to query, subset and/or aggregate HarvestChoice layers. This method may also be used to summarize classified variables along intervals of continuous variables, e.g. getLayer(var="AEZ16_CLAS", by="bmi"). The dominant class of a variable var is defined as dominant(var). Variables may also be summarized over custom geographic areas (passed as an integer array of gridcell IDs). Note that calling getLayer is equivalent to using the convenience method hcapi with the same arguments.

getLayer(var, iso3 = "SSA", by = NULL, ids = NULL, collapse = TRUE,
  as.list = FALSE)

Arguments

var

character array of variable names (all types are accepted). Use e.g. category("poverty") to search for valid variable codes).

iso3

optional array of 3-letter country or regional code(s) to filter by. Use iso to view all available codes.

by

optional character array of variables to summarize by (all types are accepted)

ids

optional array of gridcell ids to filter by (if collapse=FALSE) or summarize by (if collapse=TRUE).

collapse

if TRUE results are collapsed (summarized) over by variables. If FALSE always return all pixel values (useful for plotting and to convert the resulting table to spatial formats)

as.list

if TRUE returns a list with variable metadata instead of a simple data.table

Value

a data.table or list of var indicators aggregated by by domains

Details

Sample API call: mean body mass index and maize yield across districts in Tanzania

$ curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/hcapi/json -d '{"var" : ["AEZ8_CLAS","bana_h"], "iso3" : "CIV", "by" : ["ADM1_NAME_ALT","ELEVATION"]}' -X POST -H 'Content-Type:application/json'

See also

hcapi and getLayerWKT

Examples

# Mean body mass index and maize yield across districts in Tanzania
x <- getLayer(c("bmi", "maiz_y"), iso3="TZA", by=c("ADM1_NAME_ALT", "ADM2_NAME_ALT"))

#> Error in `[.data.frame`(x, i): undefined columns selected

x

#> Error in eval(expr, envir, enclos): object 'x' not found

# Plot results for Mara province
require(lattice)
barchart(ADM2_NAME_ALT~bmi, x[ADM1_NAME_ALT=="Mara"], col="grey90")

#> Error in barchart.formula(ADM2_NAME_ALT ~ bmi, x[ADM1_NAME_ALT == "Mara"],     col = "grey90"): object 'x' not found

# The method may be expanded to summarize classified (discrete) variables along
# intervals of continuous variables. For example the call below returns the dominant
# agro-ecological zone and total banana harvested area over Ivory Coast's provinces
# and elevation zones
x <- hcapi(c("AEZ8_CLAS", "bana_h"), iso3="CIV", by=c("ADM1_NAME_ALT", "ELEVATION"))

#> Error in `[.data.frame`(x, i): undefined columns selected

x

#> Error in eval(expr, envir, enclos): object 'x' not found