Query, subset, summarize, and download HarvestChoice indicators

Wrapper method to query, subset and/or aggregate HarvestChoice layers. This method may also be used to summarize classified variables along intervals of continuous variables, e.g. hcapi(var="AEZ16_CLAS", by="bmi"). Here AEZ16_CLAS is a classified (categorical) raster, and bmi is a continuous raster, but the request is valid. The dominant class of AEZ16_CLAS is returned along intervals of bmi. Default interval breaks are used but custom intervals may also be defined, e.g. hcapi(var="AEZ16_CLAS", by=list(bmi=c(0, 5, 10, 15, 20, 25))). The dominant class of a variable var is defined by dominant(var). Layers may also be summarized over spatial points are areas (passed as WKT representations using argument wkt). Use the format argument to control the output format (see examples below).

hcapi(var, iso3 = "SSA", by = NULL, wkt = NULL, format = NULL, ...)

Arguments

var

character array of variable codes, passed to getLayer

iso3

character array of ISO3 country or region codes, passed to getLayer

by

character array of variable codes to summarize by, passed to getLayer

wkt

WKT representation of a spatial object (points, multipoints, or polygons, multipolygons)

format

output format, one of "data.table", "csv", "tif", "dta", "asc", "grd", "rds", else "plot" to plot the rasters, or "hist" to plot histogram and univariate statistics

...

other optional arguments passed to getLayer, genFile, or to genPlot, e.g. collapse=TRUE, as.list=TRUE, dir, pal.

Value

a data.table (or other formats) of HarvestChoice indicators optionally summarized over by domains

Details

Sample API call: mean body mass index and cassava yield across provinces and districts of Tanzania

$ curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/hcapi -d '{"var" : "cass_y", "iso3" : "CIV", "format" : "tif"}' -X POST -H 'Content-Type:application/json'

/ocpu/tmp/x0bc1ac9bdf/R/.val /ocpu/tmp/x0bc1ac9bdf/stdout /ocpu/tmp/x0bc1ac9bdf/warnings /ocpu/tmp/x0bc1ac9bdf/source /ocpu/tmp/x0bc1ac9bdf/console /ocpu/tmp/x0bc1ac9bdf/info /ocpu/tmp/x0bc1ac9bdf/files/bmi--TZA.tfw /ocpu/tmp/x0bc1ac9bdf/files/bmi--TZA.tif /ocpu/tmp/x0bc1ac9bdf/files/bmi--TZA.tif.aux.xml /ocpu/tmp/x0bc1ac9bdf/files/DESCRIPTION /ocpu/tmp/x0bc1ac9bdf/files/README

GET all generated files in a ZIP archive

$ wget http://hcapi.harvestchoice.org/ocpu/tmp/x0bc1ac9bdf/zip

Sample API call: the method may be expanded to summarize classified (discrete) variables along continuous variables. For example the call below returns the dominant agro-ecological zone and average stunting in children under 5 over Ethiopia's provinces and generic elevation zones

$ curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/hcapi/json -d '{ "var" : ["AEZ8_CLAS","stunted_moderate"], "iso3" : "ETH", "by" : ["ADM1_NAME_ALT","ELEVATION"] }' -X POST -H 'Content-Type:application/json'

See also

getLayer getLayerWKT genFile

Examples

# Mean body mass index and cassava yield across provinces and districts of Tanzania
x <- hcapi(c("bmi", "cass_y"), iso3="TZA", by=c("ADM1_NAME_ALT", "ADM2_NAME_ALT"))

#> Error in `[.data.frame`(x, i): undefined columns selected

x

#> Error in eval(expr, envir, enclos): object 'x' not found

# To identify HarvestChoice indicator codes, query the entire catalog using category()
category("poverty")

#> Error in `[.data.frame`(x, i, j): could not find function "."

# Plot results for Mara province
require(lattice)
barchart(ADM2_NAME_ALT~bmi, data=x[ADM1_NAME_ALT=="Mara"], col="grey90")

#> Error in barchart.formula(ADM2_NAME_ALT ~ bmi, data = x[ADM1_NAME_ALT ==     "Mara"], col = "grey90"): object 'x' not found

# Return mean cassava yield in Ivory Coast in GeoTIFF format
r <- hcapi("cass_y", iso3="CIV", format="tif")

#> Error in `[.data.frame`(x, i): undefined columns selected

# Plot the generated GeoTIFF raster (one band only)
require(raster)
plot(raster(r[2]))

#> Error in raster(r[2]): object 'r' not found

# The method may be expanded to summarize classified (discrete) variables along continuous
# variables. For example the call below returns dominant agro-ecological zone and
# average stunting in children under 5 over Ethiopia's provinces by elevation class
x <- hcapi(c("AEZ8_CLAS", "stunted_moderate"), iso3="ETH", by=c("ADM1_NAME_ALT", "ELEVATION"))

#> Error in `[.data.frame`(x, i): undefined columns selected

x

#> Error in eval(expr, envir, enclos): object 'x' not found

# Mean harvested maize area summarized across a custom polygon
x <- hcapi(var="maiz_h", wkt="POLYGON((-16.35819663578485006 15.36599264077935345,
-15.42501860768386379 15.69472580976947462, -15.11749403024149174 14.83577785208561117,
-16.13550642453347805 14.68731771125136376, -16.35819663578485006 15.36599264077935345))")

#> Error in getLayerWKT(var, iso3 = iso3, wkt = wkt, ...): could not find function "RS.connect"

x

#> Error in eval(expr, envir, enclos): object 'x' not found