Return HarvestChoice indicator metadata

Return complete metadata records for selected HarvestChoice indicators.

indicator(q, version = NULL, as.list = FALSE, cartoCSS = FALSE)

Arguments

q

character array of pattern(s) to search against HarvestChoice indicator codes, labels, and categories. Will attempt a fuzzy match.

version

optional version filter

as.list

if TRUE return a list of indicators grouped by category

cartoCSS

also include formatted CartoCSS rules

Value

a data.table or hierarchical list of indicator metadata

Details

# API call: metadata records for all indicators matching 'cassava' or 'population'

$ curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/indicator/json \ -d '{"q" : ["cassava", "population"]}' \ -X POST -H "Content-Type:application/json"

API call: metadata records for all indicators matching 'population' in a list

$ curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/indicator/json \ -d '{"q" : "population", "as.list" : "true"}' \ -X POST -H "Content-Type:application/json"

See also

category to view HarvestChoice indicator categories

Examples

# Show complete metadata for all HarvestChoice indicators matching 'cassava' or
'population'

#> [1] "population"

indicator(c("cassava", "population"))

#> Warning: longer object length is not a multiple of shorter object length

#> Warning: longer object length is not a multiple of shorter object length

#> Warning: longer object length is not a multiple of shorter object length

#> Warning: longer object length is not a multiple of shorter object length

#> Warning: longer object length is not a multiple of shorter object length

#>                           label     code            unit       type     period
#>  1:       Irr. cassava area '05 cass_i_h              ha continuous       2005
#>  2:           Cassava prod. '05   cass_p              mt continuous       2005
#>  3:      Rfd. cassava prod. '05 cass_r_p              mt continuous       2005
#>  4: Irr. cassava val. prod. '05 cass_i_v            Int$ continuous       2005
#>  5:           Cassava yield '05   cass_y           kg/ha continuous       2005
#>  6:      Rfd. cassava yield '05 cass_r_y           kg/ha continuous       2005
#>  7:              Rural Pop. '00 PN00_RUR         persons continuous       2000
#>  8:      Rural Pop. density '05 PD05_RUR persons/sq. km. continuous circa 2005
#>  9:              Total Pop. '05 PN05_TOT         persons continuous circa 2005
#> 10:              Total Pop. '12 PN12_TOT         persons continuous       2012
#> 11:              Urban Pop. '00 PN00_URB         persons continuous       2000
#> 12:      Urban Pop. density '05 PD05_URB persons/sq. km. continuous circa 2005
#>         category         subcategory    item
#>  1:  Agriculture      Harvested Area Cassava
#>  2:  Agriculture          Production Cassava
#>  3:  Agriculture          Production Cassava
#>  4:  Agriculture Value of Production Cassava
#>  5:  Agriculture               Yield Cassava
#>  6:  Agriculture               Yield Cassava
#>  7: Demographics          Population   Rural
#>  8: Demographics          Population   Rural
#>  9: Demographics          Population   Total
#> 10: Demographics          Population   Total
#> 11: Demographics          Population   Urban
#> 12: Demographics          Population   Urban
#>                                                                                                                                                                                                                                                                          sources
#>  1:                                                                                                                                                   You, L., U. Wood-Sichra, S. Fritz, Z. Guo, L. See, and J. Koo. 2014. Spatial Production Allocation Model (SPAM) 2005 v2.0.
#>  2:                                                                                                                                                   You, L., U. Wood-Sichra, S. Fritz, Z. Guo, L. See, and J. Koo. 2014. Spatial Production Allocation Model (SPAM) 2005 v2.0.
#>  3:                                                                                                                                                   You, L., U. Wood-Sichra, S. Fritz, Z. Guo, L. See, and J. Koo. 2014. Spatial Production Allocation Model (SPAM) 2005 v2.0.
#>  4:                                                                                                                                                   You, L., U. Wood-Sichra, S. Fritz, Z. Guo, L. See, and J. Koo. 2014. Spatial Production Allocation Model (SPAM) 2005 v2.0.
#>  5:                                                                                                                                                   You, L., U. Wood-Sichra, S. Fritz, Z. Guo, L. See, and J. Koo. 2014. Spatial Production Allocation Model (SPAM) 2005 v2.0.
#>  6:                                                                                                                                                   You, L., U. Wood-Sichra, S. Fritz, Z. Guo, L. See, and J. Koo. 2014. Spatial Production Allocation Model (SPAM) 2005 v2.0.
#>  7:                                  Center for International Earth Science Information Network (CIESIN), Columbia University Earth Center; International Food Policy Research Institute (IFPRI); The World Bank; and Centro Internacional de Agricultura Tropical (CIAT), 2004.
#>  8: Calculations based on data from: Center for International Earth Science Information Network (CIESIN), Columbia University Earth Center; International Food Policy Research Institute (IFPRI); The World Bank; and Centro Internacional de Agricultura Tropical (CIAT), 2004.
#>  9: Calculations based on data from: Center for International Earth Science Information Network (CIESIN), Columbia University Earth Center; International Food Policy Research Institute (IFPRI); The World Bank; and Centro Internacional de Agricultura Tropical (CIAT), 2004.
#> 10:                                                                                                 Bright, E. A. and Rose, A. N. (2012) "LandScan Global Population Database 2012", Oak Ridge National Laboratory, July 2013. Retrieved from http://web.ornl.gov/sci/landscan/.
#> 11:                                  Center for International Earth Science Information Network (CIESIN), Columbia University Earth Center; International Food Policy Research Institute (IFPRI); The World Bank; and Centro Internacional de Agricultura Tropical (CIAT), 2004.
#> 12: Calculations based on data from: Center for International Earth Science Information Network (CIESIN), Columbia University Earth Center; International Food Policy Research Institute (IFPRI); The World Bank; and Centro Internacional de Agricultura Tropical (CIAT), 2004.
#>                     contact
#>  1: u.wood-sichra@cgiar.org
#>  2: u.wood-sichra@cgiar.org
#>  3: u.wood-sichra@cgiar.org
#>  4: u.wood-sichra@cgiar.org
#>  5: u.wood-sichra@cgiar.org
#>  6: u.wood-sichra@cgiar.org
#>  7:         z.guo@cgiar.org
#>  8:       m.bacou@cgiar.org
#>  9:       m.bacou@cgiar.org
#> 10:       m.bacou@cgiar.org
#> 11:         z.guo@cgiar.org
#> 12:       m.bacou@cgiar.org
#>                                                                                                                                                                                                                                                                                   description
#>  1:                                                                                    Cassava Irrigated Harvested Area (ha, 2005). Spatially disaggregated production statistics of circa 2005 using the Spatial Production Allocation Model (SPAM). Values are for 5 arc-minute grid cells.
#>  2:                                                                                                  Cassava Production (mt, 2005). Spatially disaggregated production statistics of circa 2005 using the Spatial Production Allocation Model (SPAM). Values are for 5 arc-minute grid cells.
#>  3:                                                                                          Cassava Rainfed Production (mt, 2005). Spatially disaggregated production statistics of circa 2005 using the Spatial Production Allocation Model (SPAM). Values are for 5 arc-minute grid cells.
#>  4:                                                                                Cassava Irrigated Value Production (Int$, 2005). Spatially disaggregated production statistics of circa 2005 using the Spatial Production Allocation Model (SPAM). Values are for 5 arc-minute grid cells.
#>  5:                                                                                                    Cassava Yield (kg/ha, 2005). Spatially disaggregated production statistics of circa 2005 using the Spatial Production Allocation Model (SPAM). Values are for 5 arc-minute grid cells.
#>  6:                                                                                            Cassava Rainfed Yield (kg/ha, 2005). Spatially disaggregated production statistics of circa 2005 using the Spatial Production Allocation Model (SPAM). Values are for 5 arc-minute grid cells.
#>  7:                                                                                                                                                                      Population, rural (2000). Rural population calculated from rural population density in each gridcell (at year 2000).
#>  8:                                                                                                                                                                                              Population Density, rural (pers./sq. km., circa 2005). Number of rural people per area unit.
#>  9: Population, total (circa 2005). Total population in 2005 derived from 29 nationally representative household surveys and population census information conducted in various years (circa 2005). Other countries 2005 population estimates derived from World Development Indicators 2008.
#> 10:                                                                                                                                                   Population, Total (2012). Total population downscaled from LandScan 2012 population counts at 30 arc-second to 5 arc-minute resolution.
#> 11:                                                                                                                                                         Population, urban (2000). Urban population calculated as difference between total population and rural population (at year 2000).
#> 12:                                                                                                                                                                                              Population Density, urban (pers./sq. km., circa 2005). Number of urban people per area unit.
#>                                                                                                                                                                                                                                                citation
#>  1:           HarvestChoice, 2015. "Cassava Irrigated Harvested Area (ha, 2005)", International Food Policy Research Institute, Washington, DC., and University of Minnesota, St. Paul, MN. Available online at http://harvestchoice.org/data/cass_i_h.
#>  2:                           HarvestChoice, 2015. "Cassava Production (mt, 2005)", International Food Policy Research Institute, Washington, DC., and University of Minnesota, St. Paul, MN. Available online at http://harvestchoice.org/data/cass_p.
#>  3:                 HarvestChoice, 2015. "Cassava Rainfed Production (mt, 2005)", International Food Policy Research Institute, Washington, DC., and University of Minnesota, St. Paul, MN. Available online at http://harvestchoice.org/data/cass_r_p.
#>  4:       HarvestChoice, 2015. "Cassava Irrigated Value Production (Int$, 2005)", International Food Policy Research Institute, Washington, DC., and University of Minnesota, St. Paul, MN. Available online at http://harvestchoice.org/data/cass_i_v.
#>  5:                             HarvestChoice, 2015. "Cassava Yield (kg/ha, 2005)", International Food Policy Research Institute, Washington, DC., and University of Minnesota, St. Paul, MN. Available online at http://harvestchoice.org/data/cass_y.
#>  6:                   HarvestChoice, 2015. "Cassava Rainfed Yield (kg/ha, 2005)", International Food Policy Research Institute, Washington, DC., and University of Minnesota, St. Paul, MN. Available online at http://harvestchoice.org/data/cass_r_y.
#>  7:                              HarvestChoice, 2015. "Population, rural (2000)", International Food Policy Research Institute, Washington, DC., and University of Minnesota, St. Paul, MN. Available online at http://harvestchoice.org/data/PN00_RUR.
#>  8: HarvestChoice, 2015. "Population Density, rural (pers./sq. km., circa 2005)", International Food Policy Research Institute, Washington, DC., and University of Minnesota, St. Paul, MN. Available online at http://harvestchoice.org/data/PD05_RUR.
#>  9:                        HarvestChoice, 2015. "Population, total (circa 2005)", International Food Policy Research Institute, Washington, DC., and University of Minnesota, St. Paul, MN. Available online at http://harvestchoice.org/data/PN05_TOT.
#> 10:                              HarvestChoice, 2015. "Population, Total (2012)", International Food Policy Research Institute, Washington, DC., and University of Minnesota, St. Paul, MN. Available online at http://harvestchoice.org/data/PN12_TOT.
#> 11:                              HarvestChoice, 2015. "Population, urban (2000)", International Food Policy Research Institute, Washington, DC., and University of Minnesota, St. Paul, MN. Available online at http://harvestchoice.org/data/PN00_URB.
#> 12: HarvestChoice, 2015. "Population Density, urban (pers./sq. km., circa 2005)", International Food Policy Research Institute, Washington, DC., and University of Minnesota, St. Paul, MN. Available online at http://harvestchoice.org/data/PD05_URB.
#>          version                                                aggFormula
#>  1:   SChEF r2.5                                 sum(cass_i_h, na.rm=TRUE)
#>  2:   SChEF r2.5                                   sum(cass_p, na.rm=TRUE)
#>  3:   SChEF r2.5                                 sum(cass_r_p, na.rm=TRUE)
#>  4:   SChEF r2.5                                 sum(cass_i_v, na.rm=TRUE)
#>  5:   SChEF r2.5      sum(cass_p, na.rm=TRUE)/sum(cass_h, na.rm=TRUE)*1000
#>  6:   SChEF r2.5  sum(cass_r_p, na.rm=TRUE)/sum(cass_r_h, na.rm=TRUE)*1000
#>  7: SChEF r12.05                                 sum(PN00_RUR, na.rm=TRUE)
#>  8:   SChEF r2.2 (sum(PN05_RUR, na.rm=TRUE)/sum(AREA_RUR, na.rm=TRUE))*100
#>  9:   SChEF r2.2                                 sum(PN05_TOT, na.rm=TRUE)
#> 10:   SChEF r2.5                                 sum(PN12_TOT, na.rm=TRUE)
#> 11: SChEF r12.05                                 sum(PN00_URB, na.rm=TRUE)
#> 12:   SChEF r2.2 (sum(PN05_URB, na.rm=TRUE)/sum(AREA_URB, na.rm=TRUE))*100
#>     isRaster                                    url
#>  1:     TRUE http://harvestchoice.org/data/cass_i_h
#>  2:     TRUE   http://harvestchoice.org/data/cass_p
#>  3:     TRUE http://harvestchoice.org/data/cass_r_p
#>  4:     TRUE http://harvestchoice.org/data/cass_i_v
#>  5:     TRUE   http://harvestchoice.org/data/cass_y
#>  6:     TRUE http://harvestchoice.org/data/cass_r_y
#>  7:     TRUE http://harvestchoice.org/data/pn00_rur
#>  8:     TRUE http://harvestchoice.org/data/pd05_rur
#>  9:     TRUE http://harvestchoice.org/data/pn05_tot
#> 10:     TRUE http://harvestchoice.org/data/pn12_tot
#> 11:     TRUE http://harvestchoice.org/data/pn00_urb
#> 12:     TRUE http://harvestchoice.org/data/pd05_urb
#>                                                                                        wms
#>  1: http://apps.harvestchoice.org/arcgis/rest/services/spam05_cell5m_h/MapServer/WMSServer
#>  2: http://apps.harvestchoice.org/arcgis/rest/services/spam05_cell5m_p/MapServer/WMSServer
#>  3: http://apps.harvestchoice.org/arcgis/rest/services/spam05_cell5m_p/MapServer/WMSServer
#>  4: http://apps.harvestchoice.org/arcgis/rest/services/spam05_cell5m_v/MapServer/WMSServer
#>  5: http://apps.harvestchoice.org/arcgis/rest/services/spam05_cell5m_y/MapServer/WMSServer
#>  6: http://apps.harvestchoice.org/arcgis/rest/services/spam05_cell5m_y/MapServer/WMSServer
#>  7:    http://apps.harvestchoice.org/arcgis/rest/services/cell5m_socio/MapServer/WMSServer
#>  8:    http://apps.harvestchoice.org/arcgis/rest/services/cell5m_socio/MapServer/WMSServer
#>  9:    http://apps.harvestchoice.org/arcgis/rest/services/cell5m_socio/MapServer/WMSServer
#> 10:    http://apps.harvestchoice.org/arcgis/rest/services/cell5m_socio/MapServer/WMSServer
#> 11:    http://apps.harvestchoice.org/arcgis/rest/services/cell5m_socio/MapServer/WMSServer
#> 12:    http://apps.harvestchoice.org/arcgis/rest/services/cell5m_socio/MapServer/WMSServer
#>     downloadedOn
#>  1:   2017-02-21
#>  2:   2017-02-21
#>  3:   2017-02-21
#>  4:   2017-02-21
#>  5:   2017-02-21
#>  6:   2017-02-21
#>  7:   2017-02-21
#>  8:   2017-02-21
#>  9:   2017-02-21
#> 10:   2017-02-21
#> 11:   2017-02-21
#> 12:   2017-02-21

# Show complete metadata for all HarvestChoice indicators matching 'population'
# in a hierarchical list
indicator("population", as.list=TRUE)[1:2]

#> $Demographics
#> $Demographics$Population
#> $Demographics$Population$Rural
#>                     label     code            unit       type     period
#> 1:         Rural Pop. '00 PN00_RUR         persons continuous       2000
#> 2:         Rural Pop. '05 PN05_RUR         persons continuous circa 2005
#> 3: Rural Pop. density '00 PD00_RUR  persons/sq. km continuous       2000
#> 4: Rural Pop. density '05 PD05_RUR persons/sq. km. continuous circa 2005
#>        category subcategory  item
#> 1: Demographics  Population Rural
#> 2: Demographics  Population Rural
#> 3: Demographics  Population Rural
#> 4: Demographics  Population Rural
#>                                                                                                                                                                                                                                                                         sources
#> 1:                                  Center for International Earth Science Information Network (CIESIN), Columbia University Earth Center; International Food Policy Research Institute (IFPRI); The World Bank; and Centro Internacional de Agricultura Tropical (CIAT), 2004.
#> 2: Calculations based on data from: Center for International Earth Science Information Network (CIESIN), Columbia University Earth Center; International Food Policy Research Institute (IFPRI); The World Bank; and Centro Internacional de Agricultura Tropical (CIAT), 2004.
#> 3:                                  Center for International Earth Science Information Network (CIESIN), Columbia University Earth Center; International Food Policy Research Institute (IFPRI); The World Bank; and Centro Internacional de Agricultura Tropical (CIAT), 2004.
#> 4: Calculations based on data from: Center for International Earth Science Information Network (CIESIN), Columbia University Earth Center; International Food Policy Research Institute (IFPRI); The World Bank; and Centro Internacional de Agricultura Tropical (CIAT), 2004.
#>                    contact
#> 1:         z.guo@cgiar.org
#> 2:       m.bacou@cgiar.org
#> 3: u.wood-sichra@cgiar.org
#> 4:       m.bacou@cgiar.org
#>                                                                                                                                                                                                                                                                                        description
#> 1:                                                                                                                                                                            Population, rural (2000). Rural population calculated from rural population density in each gridcell (at year 2000).
#> 2: Population, rural (circa 2005). Rural population in 2005 derived from 29 nationally representative household surveys and population census information conducted in various years (circa 2005). Other countries 2005 Rural population estimates derived from World Development Indicators 2008.
#> 3:                                                                                                                                                                                                           Population Density, rural (pers./sq. km, 2000). Number of Rural people per area unit.
#> 4:                                                                                                                                                                                                    Population Density, rural (pers./sq. km., circa 2005). Number of rural people per area unit.
#>                                                                                                                                                                                                                                               citation
#> 1:                              HarvestChoice, 2015. "Population, rural (2000)", International Food Policy Research Institute, Washington, DC., and University of Minnesota, St. Paul, MN. Available online at http://harvestchoice.org/data/PN00_RUR.
#> 2:                        HarvestChoice, 2015. "Population, rural (circa 2005)", International Food Policy Research Institute, Washington, DC., and University of Minnesota, St. Paul, MN. Available online at http://harvestchoice.org/data/PN05_RUR.
#> 3:        HarvestChoice, 2015. "Population Density, rural (pers./sq. km, 2000)", International Food Policy Research Institute, Washington, DC., and University of Minnesota, St. Paul, MN. Available online at http://harvestchoice.org/data/PD00_RUR.
#> 4: HarvestChoice, 2015. "Population Density, rural (pers./sq. km., circa 2005)", International Food Policy Research Institute, Washington, DC., and University of Minnesota, St. Paul, MN. Available online at http://harvestchoice.org/data/PD05_RUR.
#>         version                                                aggFormula
#> 1: SChEF r12.05                                 sum(PN00_RUR, na.rm=TRUE)
#> 2:   SChEF r2.2                                 sum(PN05_RUR, na.rm=TRUE)
#> 3: SChEF r12.05   sum(PN00_RUR, na.rm=TRUE)/sum(AREA_RUR, na.rm=TRUE)*100
#> 4:   SChEF r2.2 (sum(PN05_RUR, na.rm=TRUE)/sum(AREA_RUR, na.rm=TRUE))*100
#>    isRaster                                    url
#> 1:     TRUE http://harvestchoice.org/data/pn00_rur
#> 2:     TRUE http://harvestchoice.org/data/pn05_rur
#> 3:     TRUE http://harvestchoice.org/data/pd00_rur
#> 4:     TRUE http://harvestchoice.org/data/pd05_rur
#>                                                                                    wms
#> 1: http://apps.harvestchoice.org/arcgis/rest/services/cell5m_socio/MapServer/WMSServer
#> 2: http://apps.harvestchoice.org/arcgis/rest/services/cell5m_socio/MapServer/WMSServer
#> 3: http://apps.harvestchoice.org/arcgis/rest/services/cell5m_socio/MapServer/WMSServer
#> 4: http://apps.harvestchoice.org/arcgis/rest/services/cell5m_socio/MapServer/WMSServer
#>    downloadedOn
#> 1:   2017-02-21
#> 2:   2017-02-21
#> 3:   2017-02-21
#> 4:   2017-02-21
#> 
#> $Demographics$Population$Total
#>                   label     code            unit       type     period
#> 1:     Pop. density '05 PD05_TOT persons/sq. km. continuous circa 2005
#> 2:       Total Pop. '00 PN00_TOT         persons continuous       2000
#> 3:       Total Pop. '05 PN05_TOT         persons continuous circa 2005
#> 4:       Total Pop. '12 PN12_TOT         persons continuous       2012
#> 5: Total Pop. Dens. '12 PD12_TOT persons/sq. km. continuous       2012
#>        category subcategory  item
#> 1: Demographics  Population Total
#> 2: Demographics  Population Total
#> 3: Demographics  Population Total
#> 4: Demographics  Population Total
#> 5: Demographics  Population Total
#>                                                                                                                                                                                                                                                                         sources
#> 1: Calculations based on data from: Center for International Earth Science Information Network (CIESIN), Columbia University Earth Center; International Food Policy Research Institute (IFPRI); The World Bank; and Centro Internacional de Agricultura Tropical (CIAT), 2004.
#> 2:                                  Center for International Earth Science Information Network (CIESIN), Columbia University Earth Center; International Food Policy Research Institute (IFPRI); The World Bank; and Centro Internacional de Agricultura Tropical (CIAT), 2004.
#> 3: Calculations based on data from: Center for International Earth Science Information Network (CIESIN), Columbia University Earth Center; International Food Policy Research Institute (IFPRI); The World Bank; and Centro Internacional de Agricultura Tropical (CIAT), 2004.
#> 4:                                                                                                 Bright, E. A. and Rose, A. N. (2012) "LandScan Global Population Database 2012", Oak Ridge National Laboratory, July 2013. Retrieved from http://web.ornl.gov/sci/landscan/.
#> 5:                                                                                                 Bright, E. A. and Rose, A. N. (2012) "LandScan Global Population Database 2012", Oak Ridge National Laboratory, July 2013. Retrieved from http://web.ornl.gov/sci/landscan/.
#>              contact
#> 1: m.bacou@cgiar.org
#> 2:   z.guo@cgiar.org
#> 3: m.bacou@cgiar.org
#> 4: m.bacou@cgiar.org
#> 5: m.bacou@cgiar.org
#>                                                                                                                                                                                                                                                                                  description
#> 1:                                                                                                                                                                                                    Population Density, total (pers./sq. km., circa 2005). Number of people per area unit.
#> 2:                                                                                                                                                                            Population, total (2000). Total population calculated from population density in each gridcell (at year 2000).
#> 3: Population, total (circa 2005). Total population in 2005 derived from 29 nationally representative household surveys and population census information conducted in various years (circa 2005). Other countries 2005 population estimates derived from World Development Indicators 2008.
#> 4:                                                                                                                                                   Population, Total (2012). Total population downscaled from LandScan 2012 population counts at 30 arc-second to 5 arc-minute resolution.
#> 5:                                                                                                                                  Population Density, Total (2012). Total population density downscaled from LandScan 2012 population density at 30 arc-second to 5 arc-minute resolution.
#>                                                                                                                                                                                                                                               citation
#> 1: HarvestChoice, 2015. "Population Density, total (pers./sq. km., circa 2005)", International Food Policy Research Institute, Washington, DC., and University of Minnesota, St. Paul, MN. Available online at http://harvestchoice.org/data/PD05_TOT.
#> 2:                              HarvestChoice, 2015. "Population, total (2000)", International Food Policy Research Institute, Washington, DC., and University of Minnesota, St. Paul, MN. Available online at http://harvestchoice.org/data/PN00_TOT.
#> 3:                        HarvestChoice, 2015. "Population, total (circa 2005)", International Food Policy Research Institute, Washington, DC., and University of Minnesota, St. Paul, MN. Available online at http://harvestchoice.org/data/PN05_TOT.
#> 4:                              HarvestChoice, 2015. "Population, Total (2012)", International Food Policy Research Institute, Washington, DC., and University of Minnesota, St. Paul, MN. Available online at http://harvestchoice.org/data/PN12_TOT.
#> 5:                      HarvestChoice, 2015. "Population Density, Total (2012)", International Food Policy Research Institute, Washington, DC., and University of Minnesota, St. Paul, MN. Available online at http://harvestchoice.org/data/PD12_TOT.
#>         version                                                  aggFormula
#> 1:   SChEF r2.2 (sum(PN05_TOT, na.rm=TRUE)/sum(AREA_TOTAL, na.rm=TRUE))*100
#> 2: SChEF r12.05                                   sum(PN00_TOT, na.rm=TRUE)
#> 3:   SChEF r2.2                                   sum(PN05_TOT, na.rm=TRUE)
#> 4:   SChEF r2.5                                   sum(PN12_TOT, na.rm=TRUE)
#> 5:   SChEF r2.5       sum(PN12_TOT, na.rm=TRUE)/sum(AREA_TOTAL, na.rm=TRUE)
#>    isRaster                                    url
#> 1:     TRUE http://harvestchoice.org/data/pd05_tot
#> 2:     TRUE http://harvestchoice.org/data/pn00_tot
#> 3:     TRUE http://harvestchoice.org/data/pn05_tot
#> 4:     TRUE http://harvestchoice.org/data/pn12_tot
#> 5:     TRUE http://harvestchoice.org/data/pd12_tot
#>                                                                                    wms
#> 1: http://apps.harvestchoice.org/arcgis/rest/services/cell5m_socio/MapServer/WMSServer
#> 2: http://apps.harvestchoice.org/arcgis/rest/services/cell5m_socio/MapServer/WMSServer
#> 3: http://apps.harvestchoice.org/arcgis/rest/services/cell5m_socio/MapServer/WMSServer
#> 4: http://apps.harvestchoice.org/arcgis/rest/services/cell5m_socio/MapServer/WMSServer
#> 5: http://apps.harvestchoice.org/arcgis/rest/services/cell5m_socio/MapServer/WMSServer
#>    downloadedOn
#> 1:   2017-02-21
#> 2:   2017-02-21
#> 3:   2017-02-21
#> 4:   2017-02-21
#> 5:   2017-02-21
#> 
#> $Demographics$Population$Urban
#>                     label     code            unit       type     period
#> 1:         Urban Pop. '00 PN00_URB         persons continuous       2000
#> 2:         Urban Pop. '05 PN05_URB         persons continuous circa 2005
#> 3: Urban Pop. density '05 PD05_URB persons/sq. km. continuous circa 2005
#>        category subcategory  item
#> 1: Demographics  Population Urban
#> 2: Demographics  Population Urban
#> 3: Demographics  Population Urban
#>                                                                                                                                                                                                                                                                         sources
#> 1:                                  Center for International Earth Science Information Network (CIESIN), Columbia University Earth Center; International Food Policy Research Institute (IFPRI); The World Bank; and Centro Internacional de Agricultura Tropical (CIAT), 2004.
#> 2: Calculations based on data from: Center for International Earth Science Information Network (CIESIN), Columbia University Earth Center; International Food Policy Research Institute (IFPRI); The World Bank; and Centro Internacional de Agricultura Tropical (CIAT), 2004.
#> 3: Calculations based on data from: Center for International Earth Science Information Network (CIESIN), Columbia University Earth Center; International Food Policy Research Institute (IFPRI); The World Bank; and Centro Internacional de Agricultura Tropical (CIAT), 2004.
#>              contact
#> 1:   z.guo@cgiar.org
#> 2: m.bacou@cgiar.org
#> 3: m.bacou@cgiar.org
#>                                                                                                                                                                                                                                                                                        description
#> 1:                                                                                                                                                               Population, urban (2000). Urban population calculated as difference between total population and rural population (at year 2000).
#> 2: Population, urban (circa 2005). Urban population in 2005 derived from 29 nationally representative household surveys and population census information conducted in various years (circa 2005). Other countries 2005 urban population estimates derived from World Development Indicators 2008.
#> 3:                                                                                                                                                                                                    Population Density, urban (pers./sq. km., circa 2005). Number of urban people per area unit.
#>                                                                                                                                                                                                                                               citation
#> 1:                              HarvestChoice, 2015. "Population, urban (2000)", International Food Policy Research Institute, Washington, DC., and University of Minnesota, St. Paul, MN. Available online at http://harvestchoice.org/data/PN00_URB.
#> 2:                        HarvestChoice, 2015. "Population, urban (circa 2005)", International Food Policy Research Institute, Washington, DC., and University of Minnesota, St. Paul, MN. Available online at http://harvestchoice.org/data/PN05_URB.
#> 3: HarvestChoice, 2015. "Population Density, urban (pers./sq. km., circa 2005)", International Food Policy Research Institute, Washington, DC., and University of Minnesota, St. Paul, MN. Available online at http://harvestchoice.org/data/PD05_URB.
#>         version                                                aggFormula
#> 1: SChEF r12.05                                 sum(PN00_URB, na.rm=TRUE)
#> 2:   SChEF r2.2                                 sum(PN05_URB, na.rm=TRUE)
#> 3:   SChEF r2.2 (sum(PN05_URB, na.rm=TRUE)/sum(AREA_URB, na.rm=TRUE))*100
#>    isRaster                                    url
#> 1:     TRUE http://harvestchoice.org/data/pn00_urb
#> 2:     TRUE http://harvestchoice.org/data/pn05_urb
#> 3:     TRUE http://harvestchoice.org/data/pd05_urb
#>                                                                                    wms
#> 1: http://apps.harvestchoice.org/arcgis/rest/services/cell5m_socio/MapServer/WMSServer
#> 2: http://apps.harvestchoice.org/arcgis/rest/services/cell5m_socio/MapServer/WMSServer
#> 3: http://apps.harvestchoice.org/arcgis/rest/services/cell5m_socio/MapServer/WMSServer
#>    downloadedOn
#> 1:   2017-02-21
#> 2:   2017-02-21
#> 3:   2017-02-21
#> 
#> 
#> 
#> $<NA>
#> NULL
#>