getLayer {hcapi3}	R Documentation

Subset, and/or summarize HarvestChoice 5-arc-minute layers

Description

Workhorse method to subset and/or aggregate HarvestChoice layers. This method also aggregates classified variables by continuous variables.\ e.g. getLayer(var=c("whea_h", "AEZ16_CLAS"), by=c("ADM2_NAME_ALT", "bmi")). It does so by returning the dominant class of a classified variable within each by class, and by automatically classifying any continuous variable passed to by using default value breaks as specified in the variable metadata. The formula used to aggregate classified variables by choosing the dominant class is names(sort(table({varCode}), decreasing=T)[1]). This formula computes the frequency of each class, ranks them by decreasing frequency, and retains the top one. Layers can also be summarized over a spatial area (passed as an integer array of CELL5M ids).

Usage

getLayer(var, iso3 = "SSA", by = NULL, ids = NULL, collapse = TRUE,
  as.class = "data.table")

Arguments

var
character array of variable names (all types are accepted)

iso3
optional array of country or regional codes to filter by (3-letter code)

by
optional character array of variables to group by (all types are accepted)

ids
optional gridcell ids to return (if collapse=F) or summarize by (if collapse=T)

collapse
if FALSE always return all pixel values (useful for plotting and to convert to spatial formats)

as.class
c("data.table", "list") by default returns a simple data.table. If as.class="list" returns a well-constructed list with variable metadata

Value

a data.table (or json array) of var indicators aggregated by by domains

[Package hcapi3 version 3.5 Index]